+++ src/game/level-settings/scrambledSettings.ts
@@ -1,8 +1,8 @@
-import { LevelSettings } from "../level-creation/generateRandomLevel";
+import { SettingsProducer } from "../level-creation/generateRandomLevel";
 
 import { getNormalSettings as getNormalSettings } from "./normalSettings";
 
-export const getSettings = (difficulty: number): LevelSettings => ({
+export const getSettings: SettingsProducer = (difficulty) => ({
   ...getNormalSettings(difficulty),
   playMoves: [7, 0.3],
 });