+++ src/modules/Settings/DataTransfer.tsx
@@ -7,7 +7,6 @@ import { TransparentButton } from "@/ui/TransparentButton/TransparentButton";
 
 import info from "@/../package.json";
 import { decryptData, encryptData } from "@/support/dataTransfer";
-import { getGameValue } from "@/support/useGameStorage";
 
 import { DataFormat, getGameData, setGameData } from "./gameData";
 
@@ -56,6 +55,13 @@ const importImageData = async (
                 message: `Version mismatch: ${data.version} vs. ${info.version}`,
               };
             }
+            if (age < 0 || data.time !== data.timestamp) {
+              resolve({
+                success: false,
+                message: "Data is invalid",
+              });
+              return;
+            }
             if (age > DATA_VALIDITY_TIME) {
               resolve({
                 success: false,
@@ -66,20 +72,16 @@ const importImageData = async (
               });
               return;
             }
-            const currentLevel = (await getGameValue<number>("levelNr")) ?? 0;
             const importLevel = data.levelNr;
-
-            if (currentLevel > importLevel) {
-              if (
-                confirm("Import has less progress than current game. Continue?")
-              ) {
-                await setGameData(data);
-              } else {
-                resolve({ success: false, message: "Import canceled" });
-                return;
-              }
-            } else {
+            if (
+              confirm(
+                `You are about to import data at level ${importLevel + 1}. Continue?`
+              )
+            ) {
               await setGameData(data);
+            } else {
+              resolve({ success: false, message: "Import canceled" });
+              return;
             }
 
             resolve({ success: true, message: "" });
@@ -105,6 +107,7 @@ const DataTransfer: React.FC = () => {
   const [startDownload, setStartDownload] = useState(false);
   const [importErrors, setImportErrors] = useState<string | null>(null);
   const [importSuccess, setImportSuccess] = useState<boolean>(false);
+  const [fileInputKey, setFileInputKey] = useState(0);
 
   const encryptedData = useMemo(async () => {
     if (startDownload) {
@@ -124,20 +127,18 @@ const DataTransfer: React.FC = () => {
 
   return (
     <div className="flex flex-col gap-4">
-      <h2 className="font-bold text-lg">
-        Data transfer <sup>beta</sup>
-      </h2>
+      <h2 className="font-bold text-lg">Transfer game data</h2>
       {supportsEncryption ? (
         <>
           {!startDownload && (
             <>
               <p className="text-sm">
-                You can export your game data as an image. This image can be
-                used to import your game data into another instance of the game.
-                The export is valid for{" "}
+                You can transfer your game data to another instance of the game
+                through an image. The export is valid for{" "}
                 <strong>{DATA_VALIDITY_TIME} minutes</strong>, and needs to be
-                imported into the same version of the game.
+                imported into the same version ({info.version}) of the game.
               </p>
+              <p className="text-sm">This is not meant to act as a backup.</p>
               <TransparentButton
                 onClick={() => {
                   setStartDownload(true);
@@ -174,6 +175,7 @@ const DataTransfer: React.FC = () => {
             <input
               id="dropzone-file"
               type="file"
+              key={fileInputKey}
               className="hidden"
               onChange={async (event) => {
                 setImportSuccess(false);
@@ -186,6 +188,7 @@ const DataTransfer: React.FC = () => {
                 } else {
                   setImportErrors("No file selected");
                 }
+                setFileInputKey((v) => (v + 1) % 5);
               }}
             />
           </label>